shell: Introduce LISA_PRESERVE_SHELL
Sourcing init_env basically provides a set of bash functions (the lisa-*
commands) to the user and activates the python venv. However, that also
changes the PS1 and the colorscheme of the current shell, although this
isn't strictly necessary.

In order to give users the option of keeping their original shell
configuration, make the Lisa PS1 and colors depend on $LISA_PRESERVE_SHELL
not being set. While at it, list the env variables exported by init_env.

This results in:
  queper01@queper01-lin:~/lisa$ LISA_PRESERVE_SHELL=1 source init_env
  Activating LISA Python 3.6 venv (/data/work/lisa-next/.lisa-venv-3.6) ...
  LISA version:
      /data/work/lisa-next
      branch: heads/lisa-shell-quiet
      commit: a11a36de3e2

  Python version: 3.6

  JupyterLab v0.35.4
  No installed extensions
  Environment variables:
  EXEKALL_ARTIFACT_ROOT=/data/work/lisa-next/results
  LISA_CONF=/data/work/lisa-next/target_conf.yml
  LISA_DEVMODE=1
  LISA_HOME=/data/work/lisa-next
  LISA_HOST_ABI=x86_64
  LISA_PRESERVE_SHELL=1
  LISA_RESULT_ROOT=/data/work/lisa-next/results
  LISA_USE_VENV=1
  LISA_VENV_PATH=/data/work/lisa-next/.lisa-venv-3.6
  queper01@queper01-lin:~/lisa$

without changing colors.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>