test_series: Make match_sha1() more robust
The match_sha1() function is used to compare the kernel running on the
target with a given sha1. To do so, it issues 'uname -r' on the target
and pattern-matches to check whether the sha1 is present in the kernel
name or not. However, on targets like Pixel 3, the kernel name is
appended with a '_audio' suffix which makes the whole thing fall over.

In order to fix this, add another layer of awk filtering in match_sha1()
which removes the suffix before the pattern matching. This fixes the
issue on P3 and has no impact on existing targets with a 'clean' kernel
name.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>