lisa.conf: Add Configurable
Allow pearing a MultiSrcConf subclass with another class that is the
configured one. It will link the __init__ parameters to configuration
keys, and will provide a placeholder value in the docstring of the
class, that will contain a ":param XXX:" block for each parameter
of __init__ that is linked to a configuration key.

Also it will provide a method for __init__ parameters type checking (the
same way it is done when building the configuration instances), and will
provide a helper to map a configuration file to kwargs suitable for
__init__.

Finally, update DEFAULT_SRC class attribute with default values found in
__init__ parameters.