test_series: Change kernel flashing error to warning
The flash_sha1() functions checks if the kernel's sha1 is reported by
'uname -r' on the target to ensure the flashing procedure suceeded.
However, the sha1 may not be reported in the kernel name whenever it
matches a tag, for example.

Since this is a perfectly reasonable use-case, let's make the
'uname -r' check fail with a warning instead of an error since most
likely we can in fact proceed with the tests safely.

While at it, fix two other broken warnings.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>