trace: Introduce a TraceView
The LISA x_time_range vs TRAPpy window thing has always been a source
of confusion for me.

Ideally, I would drop whatever we have in LISA and only rely on
TRAPpy's window, but that means re-creating an FTrace instance
whenever we want to change the window. Admittedly that's not too bad
with the trace cache, but I'd like to avoid it.

The `Trace.x_{min, max}` values are currently only being *explicitely*
used by the analysis plots. `Trace.time_range` doesn't reflect the
range of this plotting window, and `Trace.start_time` doesn't reflect
the window passed to TRAPpy.

In short, it's a mess. So this is what I'm proposing instead:

- Ditch the window para for `lisa.trace.Trace`
- Introduce a TraceView class which is a view on a whole Trace that
  returns appropriate dataframe slices on df_events() calls
- Add a `lisa.trace.Trace.get_view()` interface to create TraceView
  instances
- Have `start` & `end` attributes for both Trace and TraceView