lisa/tests: RTATestBundle: Use a TraceView
Since we know the tasks that we care about, we can trim the trace
sections that come before/after those tasks were executed.

Instead of directly using the Trace in the tests, use a TraceView
trimmed down to the relevant execution window.

I've gone through the tests that implemented all or part of this logic
and replaced it with ``trace.start``/``trace.end`` where relevant.
I ran all tests and didn't see any obvious breakage.