lisa/tests: RTATestBundle: Introduce a noisy task checker decorator
Checking for noise is a common functionnality between all of the RTATestBundle
tests. However, test methods are largely free-form by design, so we do
not have any hooks into them to insert noise checking.

I see a few options to handle this:

a) Manually add the extra parameters to every test function we want
   to cover, with the relevant docstrings and result tweaks.

   This is the simplest but least maintainable solution

b) Set the thresholds as class attributes and decorate test methods,
   something like this:

   ```
   class EASBehaviour(...):
	 NOISE_THRESHOLD_PCT=5

   @classmethod
   def rta_test_method(cls, func):
       def wrapper(self, *args, **kwargs):
	   noise_res = self.test_noisy_tasks(cls.NOISE_THRESHOLD_PCT,
					     cls.NOISE_THRESHOLD_MS)

	   res = func(self, *args, **kwargs)

	   ...

   @EASBehaviour.rta_test_method
   def test_foo(self, ...):
       ...
   ```

   This is somewhat simple, and we can still tweak the class
   attributes before re-executing a test method, although that's
   not a good practice

c) What I've done, which is b) but with proper function parameters
   instead of class attributes. This involves some Python wizardry
   to expose to the user which extra parameters are exposed, but
   it is the most flexible and maintainble solution.