lisa/test: PELTTask:test_{load,util}_avg_behaviour fix
The test case uses the PELT simulator to compare against the traced PELT
signal. Due to simplifications in the PELT simulator it is not possible
to simulate exactly the same PELT signal as traced. Multiple sources of
errors commonly adds up to more than 5%. Increasing the error margin to
7% makes this problem go away.

The patch also changes the period of the rt-app task to 16.384ms to
better align with the simulated one. This allow us to rip out the time
fudging code too.

Increasing the margin does reduce the value of the test-case, but it is
either that or rewriting the PELT simulator if this test-case is to be
preserved.