lisa/test: PELTTask:test_{load,util}_avg_range remove mean check
This patch removes the mean check as it doesn't make much sense to
compare the average of signal sample value as the traced signal contains
only samples at wake/sleep and tick. Depending on when the tick happens
the average will be biases and hence a large error marging is needed.
The current error margin of 15% is actually bigger than the value range.

By dropping the mean comparison and feeding the actual duty-cycle
obtained from the traced signal to the PELT simulator to establish the
theoretical bounds of the signal, the error margins can be reduced to
1%.

Also get rid of the use of PELT simulator stableTime() as it isn't great
and we already have a constant containing the convergence time.