tools: Introduce extract-platinfo
Introduce a dummy CLI tool to get a plat_info.yml file suitable for
loading using either:
* PlatformInfo.from_yaml_map('plat_info.yml')"
* exekall run --conf plat_info.yml"

The CLI is basic as it does not allow using a different file name than
plat_info.yml, and it does not give the choice to evaluate deffered
values or not (i.e. RTapp calibration will be run, which can be lengthy
and useless depending on the use case). That would require new
Target.from_cli() API that could augment an existing
argparse.ArgumentParser() and would add more boilerplate around that
example code, so let's keep as it is for now.