lisa.analysis: Add hints for autocompletion
Since jedi autocompletion is mostly oriented toward static code
analysis, it does not pick up dynamic attributes. Add a function to give
hints to jedi, so it knows what attributes are available and their type
as well.

Alternatively, jedi can be disabled in jupyter by issuing:
%config IPCompleter.use_jedi = False
That falls back on custom ipython code that is probably less maintained
these days, so probably not as featureful.

To debug autocompletion in Jupyter notebooks:
%config IPCompleter.use_jedi = True
%config IPCompleter.debug = True