test_series: Fix USB reconnection
A typo in the command to disconnect the USB using ACME makes the command
fail, which prevents the USB reconnection to actually happen. Also, the
retry delay is sometimes too short for the device to be visible using
ADB.

In order to fix these issues, fix the typo and increase the retry delay.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>