tests: eas_behaviour: Fix TwoBigThreeSmall duty cycle
The big tasks in TwoBigThreeSmall are 80%. Because of this, EAS is kept
overutilized consistently, hence not giving the small tasks an
opportunity to be placed appropriately on little CPUs.

Although this is arguably something that could be improved on the kernel
side, it should also be noted that these tasks used to be 70% in the old
Lisa implementation, hence avoiding the problem altogether. Since the
change to 80% was not intended, and since the current test failures
don't teach us anything that we didn't know before, let's revert to 70%
tasks and avoid unnecessary pain.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>