lisa/tests: RTATestBundle: introduce test_noisy_tasks
Now that we have a proper TraceView set up for us, we can run some
analysis on the trace that won't be disturbed by setup/teardown.

A very useful test based on this is to check the "noise level" of the
workload execution, IOW test how much CPU time was used on
non-rtapp ("noisy") tasks.

The test method is not annotated on purpose so that it doesn't get
picked up by exekall.