wltests: Use repo branches instead of manifest sha1s
With repo targets, wltests checks out manifests and uses `repo sync` to
fetch the kernels that must be tested. However, `repo sync` even in
local mode, can rebase things behing your back, which messes up the
testing.

In order to fix this, switch to using proper repo branches for wltests.
The workflow for pixel 3 is as follows:

  1. Get the kernel source as explained in:
     https://source.android.com/setup/build/building-kernels#downloading

  2. Issue `repo start <branch name>`

  3. Go in the sub-trees (kernel and/or modules), hack the code and
     commit you changes.

  4. Repeat 2. and 3. with different branch names

  5. List in a file the _name_ of the repo branches you want to test
     (instead of the "<sha1> <commit title>" usual format)

  6. Run lisa-wltests-series with -k pointing at the repo top folder and
     -s pointing at the file defined in 5.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>