doc: Explain how to use repo for wltests
The 'lisa-wltest-series --help' command explains pretty well how wltest
should be used in general. However, since we introduced the notion of
'repo' targets for Pixel 3, the users should now follow two disctincts
work-flows for wltests depending on the target 'type'.

In order to clarify this a bit, try and explain how to use 'standard'
and 'repo' targets in the doc.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>