tests: sched_android: Add schedtune task placement tests
The schedtune boost is typically used on Android devices to bias the
task placement. Make sure this is working properly by checking if
top-app tasks end up on CPUs with enough capacity to meet their compute
demand.

This is not strictly matching the current android behaviour (that is,
maximize the CPU capacity for boosted tasks, ...) but should at least
make sure things work OK on performance side of things, and should be
more future proof in case of changes in android (removing fbt() for
example).

Signed-off-by: Quentin Perret <quentin.perret@arm.com>