tools: Add check-test-fix.py
Add a tool to re-run a given test against data collected using bisector,
like for the fortnightly integration. It will show a
regression/improvement table at the end to compare the reference data in
the bisector report with the new ones that were just generated. It can
be called as follow to check that a fix to a
RampDown:test_task_placement has been fixed on the first 10 available
traces:
$ check-test-fix.py -s 'RampDown:test_task_placement' bisector_report.yml.gz -oiterations=1-10 --only behaviour

Note: this will spawn one `exekall run` per CPU to speed up the
computation. That means the logs are output only when `exekall run`
command finishes to avoid interleaved output.