lisa.tests: load_tracking: Correctly join res_dir
Correctly propagating a path of the right type (ArtifactPath) for res_dir is
important to allow relocating the res_dir of the TestBundle when
deserializing its instances.

When a sub-res_dir must be created to be passed to another TestBundle,
make sure to correctly join the paths together to not loose information.