test_series: fix resolution of branch names
When branch names are used instead of SHA1 in a series file, the name is
resolved to a full SHA1 string which is longer then the 12 chars
actually returned by the for-each-ref command we use to lookup names.

The current code to resolve branch names is also a bit convoluted when
it comes to check that a specified SHA1 is unique.

Cleanup the branch resolution and SHA1 consistency check code.
Always use full length SHA1.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>