wltests/wa_results_collector: add sorted plot_cdf
The CDF plot reports plots in a random dataframe way and set the plot
title based on the last CDF plot added to the image.

Encode the "below threshold" percentage in the legend to report the
actual value for each test. Move the legend below the plot to ensure
optimal visibility also for big datasets. Finally, ensure to plot the
CDFs from the best to the worst thus making the legend even more useful.

An optional parameter is added to specify the number of columns in the
legend. Some more advanced code could try to figure out the best value
based on labels length but, let's keep the code simple for the time
being.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>