lisa.trace: Make events checking optional in requires_events
Using requires_events(check=False) decorator will wrap the
method/function with a checker that will not check for the available
events. That is still useful for documentation purposes, to support
automatic creation of FtraceConf and to allow chaining of checkers even
on functions/methods that do not operate on objects with a `self.trace`
attribute.