load_tracking: Expose InvarianceItem.from_target
Allows specifying a given pair of CPU and freq to run the test.
Since "int" annotations can usually not be satisfied without --sweep,
exekall will not pick it up, as expected.

To run the test directly from the InvarianceItem (and thus skipping
Invariance class):

$ lisa-test 'InvarianceItem*' --param '*InvarianceItem.from_target' 'cpu' 0 --param '*InvarianceItem.from_target' 'freq' 999000

Using the same start and stop value will make it sweep with one value
only.