lisa.analysis.task: Add TaskAnalysis.cpus_of_tasks()
Allow getting the CPUs on which the given tasks executed. This is useful
to convert the tasks names of the rtapp profile into list of CPUs to
feed to plotting functions that need to know what CPU to look at. Since
the trace is usually already cropped around the execution of rtapp, this
allows spotting the relevant pieces of data.