analysis.rta: add df getter and plotter for activation events
This adds a new pair of dataframe getter and plotter to get information
about the activations of a specified task.

The returned dataframe includes slack and latency information for each
activation. The plotter is compliant to other analysis plots and generate
activation bands which can be over-imposed to other plots.

For both methods, do allow to defined a star_time to align the generated
data/plots with other events, for example from a trace.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>