utils: Add update_wrapper_doc() helper
Behaves like functools.wraps(), with the additional capability of
updating the docstring and the signature of the wrapped function.

This allows taking into account an "extender" function that is used to
implement the wrapper, so it can add some of its parameters to the
decorated function, and part of its docstring as well.