workload: add timeout to avoid broken workloads
It may happen that, when executing a workload, it gets stuck (e.g.,
because of a deadlock, a missing wakeup signal, an unlucky infinite loop
execution branch, etc.).
devlib's Target::execute, allows to specify a timeout parameter that
generates an exception if the timeout time expires.

Expose the timeout parameter to Workload::run() function that will be
directly passed to Target::execute and forwards the exception to the
caller.

Signed-off-by: Alessio Balsini <balsini@android.com>