wlgen/workload: Keep all assets into a dedicated folder
RTApp requires a JSON description and generates log files.

Let's ensure all these assets are confined withing a dedicated folder so that
when we pull stuff we are sure to not get leftovers from previous executions.
This allows also to do a proper cleanup of the target once a test has completed.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>