lisa_shell: Change priority of folders in PATH
Give priority to LISA tools over the systems ones, since we usually ship
a version of the tools that is known to work and is newer than what
comes from the distribution.

This mainly avoids issues around trace-cmd, which is shipped in a
version that segfaults on some versions of Ubuntu with recent trace
files.