lisa.trace: Trim down attributes of Trace
Remove the following attributes:

* trace_format: client code is not supposed to change its behavior
depending on the format of the trace, as the API is fully unified. For
live debugging, type(trace.ftrace) will give that information.

* normalize_time: client code is usually written with the assumption
that the time is normalized or not, depending on how it built the trace.
No use of that attribute was found in the current code base.

* trappy_cls: unset and unused attribute, probably left over from a
previous cleanup.

* overutilized_time/overutilized_prc: status analysis now provides the
feature.

* freq_coherency: has been moved to frequency analysis, where it is
actually used.

* Deprecate Trace.ftrace attribute so that all the information that
client code needs is exposed in Trace directly. FTrace underlying object
is now an implementation detail.