tools/kmodules/sched_tp: add a patch version of the module
Which provides an alternative way for the user to get the required sched
events in a modern kernel (5.3 and above) to run Lisa tests.

The user is expected to use git am to apply the patch to their tree if
they rather this option to building an external kernel module.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>