lisa.trace: Move function stats features to lisa.analysis.functions
Follow the lisa.analysis.rta.PerfAnalysis pattern, which is to keep
things not based on events out of lisa.trace.Trace. This allows that
class to focus on event management.

Since function stats are not actually sharing any code with the other
class, split it. That is not surprising as this function stats
processing is not event-based at all.