tools: Move target binaries inside lisa package
In order to keep a filename-based API in LISA, we need the binaries to be
available along the `lisa` Python packages. The recommended way to do that is
either to use the the ResourceManager API (which would require some changes), or
to just rely on module-relative path using __file__ [1]. In the latter case, we
need to move the resources along the sources, which is what this commit is
doing.

[1] https://setuptools.readthedocs.io/en/latest/setuptools.html#accessing-data-files-at-runtime