lisa.wlgen.rta: Fix race condition in RTA workload execution
The rtapp JSON config file is pushed to the target in RTA.__init__, but is only
used when RTA.run() is called. Since the name is not unique, something else
could have modified the file in between, leading to the wrong workload to be
executed.

Fix that by prepending a UUID to the remote filename, so that there is a
one-to-one mapping between remote conf files and RTA() instances.