lisa.tests.base: Add SimpleRTATestBundle
Having RTATestBundle._from_target() implemented in RTATestBundle breaks the
generality of RTATestBundle.from_target(), as it gains parameters specific to
RTATestBundle._from_target that will be expected to be available on
implementations of _from_target in all subclasses.

Splitting that feature in SimpleRTATestBundle allows clearing the way to
TestBundle.from_target(), which does not add unnecessary constraints on the
expected signature of downstream _from_target.