lisa/wa_results_collector: Fix parse_traces=True
Daniel pointed out that passing `parse_traces=True` to
WaResultsCollector breaks horribly on sad piece of code.

We're attempting to create a PlatformInfo to give to Trace, but we're
only filling it with the kernel version which mostly useless. Just skip
that step, and only pass a non-None platinfo if the user specifed one.

Ideally we should collect one platform_info per wa directory, but that
would mean adding a LISA-specific step to lisa-wltest which I think we
all would like to avoid.