exekall: Fix expressions building
When a parameter can be built with an operator, but nothing can provide the
necessary prerequisites for that operator, the expression was unconditionally
discarded. If the parameter has a default value, it should just be ignored in
that case, and the expression should still be built.

Sanitize the expression building function and cleanly deal with these
conditions.