lisa.tests: Ban usage of rtapp_profile.keys() to filter traces
In recent versions of rtapp, the name a task will get during execution is not
the one specified, in order to distinguish between forked tasks.
`self.rtapp_tasks` is a property of RTATestBundle that allows getting the names
as they appear in the trace, which is therefore suitable for filtering.