lisa.conf: Add KeyDesc.__init__(newtype='Foo') parameter
This allows creating a type for the key, that can then be used in annotations.
This type is set as an attribute of the configuration class for scoping. The
configuration class will also have a properly annotated getter so that exekall
will know how to provide values from the configuration for that type.