exekall: Use return annotation of functions to infer type when reloading from ValueDB
Using the return annotation of the producer callable allows to use "phantom"
types, with functions returning instances of something else. Once reloaded from
the ValueDB, the FrozenExprVal will actually still keep this phantom type,
rather than being "degraded" to the actual type, unless that type is a subclass
of the return annotation.