lisa/tests: misfit: Fix rt-app naming issue
Rt-app task naming was changed with the introduction of the fork
functionality, and this was reflected in

  5b066eef0578 ("lisa.tests.base: Introduce RTATestBundle.rtapp_tasks property")

The misfit test case uses the name of the task to figure out on which
CPU it is initially pinned, which is somewhat icky (and now broken with
the task naming change). Look at the full fledged profile to figure this
out instead.