lisa.tests.scheduler.misfit: Remove __init__ and dispatch in properties
Doing work in __init__ means it can raise exceptions, leading to no object being
created, which prevents easy reproduction of the issue by re-running the test on
the serialized objects (since there are no objects).

Instead, dispatch the attributes in properties that are computed as needed by
the test methods.