lisa.tests.scheduler.load_tracking: Make PELTTask.test_(util|load)_avg_behaviour more robust
Take into account differences between rt-app calibration ratios and CPU
capacities when assessing the mean of the settled signal. This is needed to
accommodate different instruction mix in the busy loop of rt-app compared to the
benchmark that was used to establish the CPU capacities.

For now, this does not have a big impact since the workload only runs on the big
CPUs, but it will becomes very important when it running on small ones.