lisa.tests.scheduler.load_tracking: Make PELT behaviour tests more robust
Since the PELT mean is not equal to the duty cycle of a task (PELT does not
decay when the task sleeps, it decays "all at once" at the next wakeup), compare
the "tunnel mean" of PELT and the duty cycle. This takes the average between the
mean of local maximums and local minimums, which is supposed to match the duty
cycle.