lisa.wlgen.rta: Allow JSON files outside of the res_dir
The way the RTA class is designed is that its base constructor is very
simple - it only really requires a target and an rt-app JSON
file. Alternative constructors then come in handy to create the JSON for
us.

I wanted to try out an rt-app JSON handed out to me by someone else, but
wlgen complains when trying to join res_dir and the jsson_file. The
docstring of the class does say it is a path to a file, not a name, so make
it so.