lisa.trace: Add FtraceCollector(trace_clock='global', saved_cmdlines_nr=8192)
Allow changing the number of entries in the cmdlines/PID table created by the
kernel while tracing. This is needed on systems generating a lot of
tasks (Android), to avoid having placeholder task names such as "<...>".

Also allow changing the trace clock source, and default to the "global"
monotonic shared clock so that timestamp between CPUs are consistent.