lisa.tests.base: Do not deepcopy the instance in TestBundle.from_target
Serializing and deserializing an object is the equivalent of a deepcopy.
Since TestBundle references PlatformInfo objects, which can be quite big
(content of kallsyms and kernel config), this would imply keeping a copy
of that info in each instance we manipulate.

To avoid that issue, use the original bundle that references a shared
PlatformInfo (the one from the Target object).