lisa.analysis.rta: Validate trace in rtapp_tasks property
Make sure the trace contains at least one of the event we look at to
detect rtapp tasks, and raise an exception if no relevant event is
available. This avoids returning an empty list in cases where the trace
could very well contain rtapp tasks, but just did not record/parse the
events.