lisa.tests.scheduler.misfit: Use rt-app phase start rather than sched_switch
Using sched_switch events to figure out when the tasks started running does
work of the time, sadly this isn't the case here: we only care about what
happens to the tasks once they all start running on their dedicated
CPUs (second phase of the rt-app profile). Instead of playing games with
time margins, use the phase start event directly.

Find the last start of the second phase, and use that as the test start
time. Those phase start events occur after the task's switch-in, which is
fine because we use squash_df: the switch-in will be brought forward to
whenever we slice the trace.