lisa.datautils: Change *_refit_index(method='inclusive')
"inclusive" windowing method is more adapted to *_refit_index, since it
will guarantee that a refitted signal will have correct values inside
the window. With 'pre', it was not the case on the right side of the
window as the last row is "forward propagated", thereby delaying the
point at which the signal takes this value.