lisa.wlgen.rta: Isolate RTapp CPU capacity computation
Make RTA.get_cpu_capacities_from_calibrations() return capacities suitable for
writing in sysfs, i.e. the capa is the same for all CPUs of the same class and
the max capacity is 1024.

This makes the warning slightly less informative for debugging since all CPUs in
the same class will be treated the same, but at the same time that's reflecting
more closely what is really going to happen when the new capacities are written
to sysfs.