lisa.platforms.platinfo: Change cpu-capacities key
BREAKING CHANGE: The key has now multiple subkeys:

cpu-capacities:
    orig: {CPU: capa, ...} # Previously "cpu-capacities" key
    rtapp: {CPU: capa, ...} # CPU capacities adjusted to fit rtapp calibration
    writeable: True/False # True if CPU capacities are writeable in sysfs

When using the capacities to create an rtapp workload in implementations of
RTATestBundle.get_rtapp_profile(), "rtapp" subkey must be used in order to work
with the capacities that will be written in sysfs by RTATestBundle.run_rtapp().

If the capacities are not writeable, the "rtapp" subkey is equal to the "orig"
subkey.