DO NOT MERGE YET
lisa.pelt: Make kernel_util_mean() accurate for non-50% tasks

Compute the "true" mean of the util_avg signal as output by the kernel. This
takes into account the fact that the util is not decayed while tasks are
sleeping. Instead, the value drops at next wakeup, which prevents from using a
simple series_mean() on the signal.