lisa.wlgen: Make Workload 'name' parameter optional
BREAKING CHANGE

The "name" parameter has changed position to make it optional. This code:

    wload = RTA.by_profile(target, "4_50_tasks", profile)

Needs to be rewritten as:

    wload = RTA.by_profile(target, profile, name="4_50_tasks")

Being forced to specify a meaningful name always buggered me, especially
since it is not at all required under the hood - we get uniquely identified
directories and all that.

That plus the qualname ought to be good enough. I was actually tempted to
remove it alltogether, but then there's things like naming the JSON file
for the rt-app calibration that makes sense.

Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>