lisa.tests.base: Avoid spurious failure of DmesgTestBundle.test_dmesg()
Tests using the performance cpufreq governor while EAS is enabled will
receive a dmesg warning like this one:

    dmesg output: kern: warn: [1602.006703] rd 0-5: Disabling EAS, schedutil is mandatory

Since it's an expected warning that is not reflecting an actual issue,
allow tests to easily ignore this warning.

Also, on kernel >= 5.6, executable stack will trigger this warning:

    kern: warn: [555.927466] process 'root/devlib-target/bin/busybox' started with executable stack

Since this is unrelated to the sort of issue we are looking for, ignore
it.