lisa.energy_model: Warn when a CPU has no idle states
Warn about missing idle states in EnergyModel.from_target() so that we
report missing idle states in the context of a specific target.

This is to be preferred to a warning in __init__ that would trigger on
any object built, even in cases where it could be legitimate, since that
is just a plain mutable data structure.