lisa.analysis.base: Fix plot_method() positional parameters
Passing positional parameters to plot methods when they have a default
values leads to passing it twice to the method, once as a positional and
then again as a keyword.

Solve the issue by factoring the positional parameters in the keyword
arguments, which simplifies the task and avoids that issue.