energy_model: fix single CPU cluster energy estimate
Some topologies like the Pixel4 have a single CPU on a cluster. In this
case, the CPU and the cluster are described similarly:

e.g. multiple CPUs cluster:

 Single CPU description: (4, )
 Cluster description: (4, 5, 6)

e.g. single CPU cluster:

 Single CPU description: (7, )
 Cluster description: (7, )

This is problematic in _estimate_from_active_time(), as the energy computed
would be overwritten. Ensure a single result table initialization when
several nodes have the same description. This avoid losing the energy
computed for the first nodes of an identical description.

Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>