exekall: include inherited methods for selected classes
Only the operators defined in the modules pointed at by the user are
used for the root operator of the selected expressions. Classes defined
in these modules are included, along with the methods they define, but
inherited methods are not included if they are not defined in the
selected modules.

Change that behaviour to include the inherited methods, since it's a
more expected behaviour.