lisa.tests: misfit: Better tie allowed idle time with CONFIG_HZ
While the load balance interval for a sched_domain is indeed built as

  1 ms * sched_domain_span(sd)

this actually ends up being translated as a jiffy multiple. For instance,
on the TC2 (HZ_100), the default balance_interval for the highest domain is
5ms, but one jiffy is 10ms, so the interval effectively becomes 10ms.

Reflect this in the default allowed idle time for misfit tasks, as it can
take up to a full balance interval before a task is migrated over.