lisa.tests.scheduler.util_tracking: Check util against correct phase
The code currently looks for the first util signal update after every
task activation and checks the values are consistent with the
expectations, based on rtapp phase number.

The issue is that the following sequence can happen:

signal update
sleep
wakeup
start of new phase
signal update

In that case, we end up checking the values from the 2nd signal update
against the previous rtapp phase number.

Fix that by resolving the rtapp phase number of the phase the signal
update is in, instead of the phase number the activation is in.