lisa.tests.base: Limit memory consumption of FtraceTestBundle.trace property
When exekall run -n is used, a lot of live FtraceTestBundle are created,
leading to excessive memory consumptions. Crank down the limits so that
dataframes are evicted early even if the parent FtraceTestBundle is kept
around for serialization.

Also, disable the on-disk parquet swap since it seems to be leaking
memory.