lisa.trace: Use HRTxtTraceParser by default for .txt trace
BREAKING CHANGE

Since text trace are typically in the human readable format (as opposed
to the semi-machine readable raw output of trace-cmd report), use this
by default for .txt extension.

If the old behaviour is still wanted, traces can be built with:

    from lisa.trace import Trace, TxtTraceParser
    Trace(..., parser=TxtTraceParser.from_txt_file)

The new behaviour is equivalent to:

    from lisa.trace import Trace, HRTxtTraceParser
    Trace(..., parser=HRTxtTraceParser.from_txt_file)

As a side effect, this will enable lisa-plot to work out of the box (to
some extent) with such text trace.