lisa.trace: Ensure meta event source df use bytes
Meta events (like userspace@... or trace_printk@...) are extracted from
existing dataframes. The text data is expected to be encoded as bytes,
so that parser that can provide bytes out of the box avoid a conversion
(and x4 in memory consumption incurred by str).

That said, some parsers may return string columns, in which case it
needs to be converted, as regexes are written against bytes type.