lisa.trace: Rename 'func' column to 'ip' for print events
BREAKING CHANGE

print/bprint/bputs events will now have an "ip" column instead of "func"
to align with kernel names, so it is consistent across parsers.

Note: There is no guarantee on whether the content is an address or
a function name anyway, so filtering it using
FunctionsAnalysis.df_resolve_ksym() is the only reliable way of getting
a function name out of it.

Note 2: This does not affect meta events like trace_printk@... or
userspace@..., which is the typical way of interacting with
print/bprint/bputs.