lisa.datautils: Add {df,series}_dereference()
Allow using a series/column as a pointer into another dataframe's
columns.

For example, one might have a dataframe with a column for the frequency
of each CPU, and a series of CPUs where a task is scheduled. This
function would allow getting the current frequency the task was running
at.