base: Print out info about the total noise
When checking for noise in the experiment to verify there was minimal
interference, we do the check against the total noise in the system but
only print out the info about the noisiest task.

In latest integration branch, RampDown:test_task_placement is undecided
in one case although the noisiest task is 0.24% and the threshold is 1%.
This pattern is repeated, hence we need to print out more info about the
total noise to help with the diagnosis.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>